create createViewModelFactory() and use it in FieldsViewModel
