+++ app/src/main/java/gr/unimatch/android/viewmodel/FieldsViewModel.kt
@@ -3,9 +3,7 @@ package gr.unimatch.android.viewmodel
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.createSavedStateHandle
-import androidx.lifecycle.viewmodel.CreationExtras
-import gr.unimatch.android.UniMatchApplication
+import gr.unimatch.android.common.createViewModelFactory
 import gr.unimatch.android.repository.FieldsRepository
 
 class FieldsViewModel(
@@ -15,21 +13,12 @@ class FieldsViewModel(
     val fields = fieldsRepository.getFields()
 
     companion object {
-        val Factory: ViewModelProvider.Factory = object : ViewModelProvider.Factory {
-            @Suppress("UNCHECKED_CAST")
-            override fun <T : ViewModel> create(
-                modelClass: Class<T>,
-                extras: CreationExtras,
-            ): T {
-                val application =
-                    checkNotNull(extras[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY])
-                val savedStateHandle = extras.createSavedStateHandle()
-
-                return FieldsViewModel(
+        val Factory: ViewModelProvider.Factory =
+            createViewModelFactory { savedStateHandle, appContainer ->
+                FieldsViewModel(
                     savedState = savedStateHandle,
-                    fieldsRepository = (application as UniMatchApplication).appContainer.fieldsRepository,
-                ) as T
+                    fieldsRepository = appContainer.fieldsRepository,
+                )
             }
-        }
     }
 }
